bsc_api: Change the cause to a uint8_t for now
The RR Cause can only be one byte. Change from uint16_t to uint8_t for that.

+++ openbsc/include/openbsc/bsc_api.h
@@ -18,7 +18,7 @@ struct bsc_api {
 			struct msgb *msg, uint16_t chosen_channel); 
 	void (*dtap)(struct gsm_subscriber_connection *conn, struct msgb *msg);
 	void (*assign_compl)(struct gsm_subscriber_connection *conn,
-			  uint16_t rr_cause);
+			  uint8_t rr_cause);
 	void (*assign_fail)(struct gsm_subscriber_connection *conn,
 			 uint16_t rr_cause);
 	int (*clear_request)(struct gsm_subscriber_connection *conn,
